+++ packages/ub/CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
  - data storage entities for "Many" attributes now added by metadata transformation hook of @unitybase/ub model (instead native server code)
- - mixin what implements data storage for ACL RLS implementation moved from native to JS (@unitybase/ub/mixins/aclRlsStorageMixin.js)
+ - mixin what implements data storage for ACL RLS implementation moved from native to JS (@unitybase/ub/mixins/aclRlsStorageMixin.js).
+  **BREAKINGS**
+   - check constraints about "all onEntity attributes is not null" (CHK_*_SOMENOTNULL and CHK_*_NOALLNOTNULL) are removed - this verified by aclRlsStorage mixin
+   -  
 
 ### Changed
  - Add `TenantUser` built-in role on login of every user, if user is logged in any tenant except